More promising solution.
Still need to sort out the UART inputs. However: 
We get 8 bits. The first bit is active high and used to clear the registers (if first bit is 1 -> clear both registers).
The second bit we send (data_in(6)) is used to either read or write to a specific register (which is specified by the next two bits).